requirements.txt
Flask==1.1.1
Werkzeug==1.0.0
requests
requests-toolbelt
google-cloud-datastore==1.10.0
google-cloud-tasks
googleapis_common_protos
